Split gen:model timestamp,soft-deletes use two options
